// frontend/src/api/generated/models/assetResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface AssetResponse {
  id: string;
  name: string;
  category: string;
  acquisition_value: string;
  acquisition_date: string;
  useful_life_months: number;
  depreciation_start_date: string;
  status: string;
  location: string | null;
}